Restructured, added editorconfig, removed .idea
