+++ .editorconfig
@@ -1,0 +1,5 @@
+root = true
+
+[*.js]
+indent_style = space
+indent_size = 4
