+++ .gitignore
@@ -1,3 +1,3 @@
 npm-debug.log*
 node_modules
-.idea/workspace.xml
+.idea
